signal: Do not allow weak connections with no strong reference
When connecting a receiver to a signal using a Weak that has no strong
references, we can call emit, which would remove this weak pointer.

This is problematic when connecting from within Rc::new_cyclic. We don't
want to accidentally remove items that are about to be constructed.

To solve this, we assert that there exists at least one strong reference
at the time of connections.