Make signal use fractional index based ordering
Instead of using strings for creating orders of signals, we use relative
orderings using fractional indexes. This is more desirable because
strings require global reasoning, and can invalidate code that is
logically "far away" from code that is currently being considered;
adding a connection with the same name may collide with some other
connection.

This is implemented by adding two methods to the Signal struct: `before`
and `after`.

These methods return a new signal to which all connections will be
ordered according to how the signal was created.

For instance:

        signal.connect(...);
        signal.before().unwrap().connect(...);

The second connection here will run before the first, once `emit` is
called.

These properties are nestable.

        /*1*/ signal.before().unwrap().connect(...);
        /*2*/ signal.before().unwrap().after().unwrap().connect(...);
        /*3*/ signal.connect(...);

The signals above, no matter their order of connection will be emitted
to according to the given comments. Any amount of `after` on a `before`
signal will never be ordered after connection 3 in the above example.

To visualize this, a signal when created using Signal::new(), will have
a fractional order of 0.5. Calling `before` returns a signal with a
fractional order of 0.25, calling `after` on that signal returns 0.375,
and so on. We don't use floats but instead use u128.

Fractional indexing is not infinite (although it could be) because of
the fixed fractional order size. That's the reason why `before` and
`after` returns an Option. If the `u128` is exhausted, we return None.

The alternative is to allocate multiple `u128` whenever one is
exhausted, but we want to avoid signals being created in loops causing
memory leaks.

Because this patch removes strings as an index, a new method for
disconnecting has been introduced: the Connection struct. Every call to
`connect` will return a `Connection`. This type exposes a method that
allows for precise disconnection.