Make after/before panic if exhausted
This is more practical than having to unwrap on each call. If the space
is exhausted, there's not much else the program can do to sort itself
out. We might add the unwrapping versions in the future if there's a
need.

This commit also adds quickcheck to ensure that orders are correctly
assigned based on after/before sequences.

We also move from binary_search_by to partition_point. The former
doesn't guarantee to return an index past the end of the partition,
whereas the latter does.