Make before/after return the same subsignal
Whenever we have a sequence of before/after calls, eventually we'll
exhaust fractional index space. Once this is exhausted, we instead
create a subsignal at that point which restarts fractional indexing.

The problem is that this exhaustion, if performed once more, should
return the exact same subsignal. This was not the case previously.
Certain undesirable effects can be produced this way.

For instance, say we run `signal.before().before()...` twice until
non-exhaustion. In that case, the `Signal` returned is still the same
signal object, at the same fractional index. Two connects to these
signals will order as expected.

If however, the sequence results in a subsignal, then there are now two
subsignals at the exhausted index. Connecting to the second object will
then be ordered after any connections to the first. This is surprising
for any user.

So instead, this commit makes sure to store the subsignals that were
used for before/after sequences, and if we already have one at a
particular exhausted fractional index, we then just return a clone to
that already-created subsignal.

This way, behavior remains stable regardless of how many nestings we
have of before/after calls.